+++ src/main/java/com/abc/Account.java
@@ -3,17 +3,12 @@ package com.abc;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Account {
+public abstract class Account {
 
-    public enum AccountType {
-        CHECKING, SAVINGS, MAXI_SAVINGS//, SUPER_SAVINGS
-    }
 
-    private final AccountType accountType;
     public List<Transaction> transactions;
 
-    public Account(AccountType accountType) {
-        this.accountType = accountType;
+    public Account() {
         this.transactions = new ArrayList<Transaction>();
     }
 
@@ -25,35 +20,17 @@ public class Account {
         }
     }
 
-public void withdraw(double amount) {
-    if (amount <= 0) {
-        throw new IllegalArgumentException("amount must be greater than zero");
-    } else {
-        transactions.add(new Transaction(-amount));
+    public void withdraw(double amount) {
+    	if (amount <= 0) {
+    		throw new IllegalArgumentException("amount must be greater than zero");
+    	} else {
+    		transactions.add(new Transaction(-amount));
+    	}
     }
-}
 
-    public double interestEarned() {
-        double amount = sumTransactions();
-        switch(accountType){
-            case SAVINGS:
-                if (amount <= 1000)
-                    return amount * 0.001;
-                else
-                    return 1 + (amount-1000) * 0.002;
-//            case SUPER_SAVINGS:
-//                if (amount <= 4000)
-//                    return 20;
-            case MAXI_SAVINGS:
-                if (amount <= 1000)
-                    return amount * 0.02;
-                if (amount <= 2000)
-                    return 20 + (amount-1000) * 0.05;
-                return 70 + (amount-2000) * 0.1;
-            default:
-                return amount * 0.001;
-        }
-    }
+    public abstract String getPrettyName();
+    
+    public abstract double interestEarned();
 
     public double sumTransactions() {
        return checkIfTransactionsExist(true);
@@ -64,10 +41,6 @@ public void withdraw(double amount) {
         for (Transaction t: transactions)
             amount += t.amount;
         return amount;
-    }
-
-    public AccountType getAccountType() {
-        return accountType;
     }
 
 }
