+++ src/main/java/com/abc/AccountChecking.java
@@ -1,0 +1,16 @@
+package com.abc;
+
+public class AccountChecking extends Account {
+
+	@Override
+	public String getPrettyName() {
+		return "Checking Account";
+	}
+
+	@Override
+	public double interestEarned() {
+		double amount = sumTransactions();
+		return amount * 0.001;
+	}
+
+}
