+++ src/main/java/com/abc/AccountMaxiSavings.java
@@ -1,0 +1,20 @@
+package com.abc;
+
+public class AccountMaxiSavings extends Account {
+
+	@Override
+	public String getPrettyName() {
+		return "Maxi Savings Account";
+	}
+
+	@Override
+	public double interestEarned() {
+		double amount = sumTransactions();
+		if (amount <= 1000)
+            return amount * 0.02;
+        if (amount <= 2000)
+            return 20 + (amount-1000) * 0.05;
+        return 70 + (amount-2000) * 0.1;
+	}
+
+}
