+++ src/main/java/com/abc/AccountSavings.java
@@ -1,0 +1,19 @@
+package com.abc;
+
+public class AccountSavings extends Account {
+
+	@Override
+	public double interestEarned() {
+		double amount = sumTransactions();
+		if (amount <= 1000)
+            return amount * 0.001;
+        else
+            return 1 + (amount-1000) * 0.002;
+	}
+
+	@Override
+	public String getPrettyName() {
+		return "Savings Account";
+	}
+
+}
