+++ src/test/java/com/abc/CustomerTest.java
@@ -10,8 +10,8 @@ public class CustomerTest {
     @Test //Test customer statement generation
     public void testApp(){
 
-        Account checkingAccount = new Account(Account.AccountType.CHECKING);
-        Account savingsAccount = new Account(Account.AccountType.SAVINGS);
+        Account checkingAccount = new AccountChecking();
+        Account savingsAccount = new AccountSavings();
 
         Customer henry = new Customer("Henry").openAccount(checkingAccount).openAccount(savingsAccount);
 
@@ -35,23 +35,23 @@ public class CustomerTest {
 
     @Test
     public void testOneAccount(){
-        Customer oscar = new Customer("Oscar").openAccount(new Account(Account.AccountType.SAVINGS));
+        Customer oscar = new Customer("Oscar").openAccount(new AccountSavings());
         assertEquals(1, oscar.getNumberOfAccounts());
     }
 
     @Test
     public void testTwoAccount(){
         Customer oscar = new Customer("Oscar")
-                .openAccount(new Account(Account.AccountType.SAVINGS));
-        oscar.openAccount(new Account(Account.AccountType.CHECKING));
+                .openAccount(new AccountSavings());
+        oscar.openAccount(new AccountChecking());
         assertEquals(2, oscar.getNumberOfAccounts());
     }
 
     @Ignore
     public void testThreeAcounts() {
         Customer oscar = new Customer("Oscar")
-                .openAccount(new Account(Account.AccountType.SAVINGS));
-        oscar.openAccount(new Account(Account.AccountType.CHECKING));
+                .openAccount(new AccountSavings());
+        oscar.openAccount(new AccountChecking());
         assertEquals(3, oscar.getNumberOfAccounts());
     }
 }
